pyproject.toml
[project]
name = "dotfiles-maintainer"
version = "1.0.0"
description = "An MCP server for managing dotfiles with semantic memory and context awareness."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Ahmad Mollaei", email = "The.AhmadMollaei@proton.me"}]
keywords = ["mcp", "dotfiles", "semantic-memory", "configuration"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "mcp>=1.25,<2",
    "google-genai>=1.57.0",
    "mem0ai>=1.0.2",
    "sentence-transformers>=5.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=0.1.2",
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "detect-secrets>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/lonlydwolf/dotfiles-maintainer"
Repository = "https://github.com/lonlydwolf/dotfiles-maintainer"
Issues = "https://github.com/lonlydwolf/dotfiles-maintainer/issues"

[project.scripts]
dotfiles-mcp = "dotfiles_maintainer.server:main"
dotfiles-cli = "dotfiles_maintainer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.basedpyright]
# typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
# Disable specific annoying rules globally
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnusedCallResult = false

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
